Use real Futures, widen APIs to pass an implicit ExecutionContext.
Currently I don't think this provides much benefit for the JES backend.  A lot of potentially blocking
code (including explicitly Thread.sleeping backoffs) still runs on the Akka ExecutionContext.  This could
be further refactored to have the execute/resume return more quickly with a different implementation of
ExecutionHandle, letting the CallExecutionActor drive the backoff to eliminate the Thread.sleeps.