Make find ' ' in filename safe (#3864)
* Make find ' ' in filename safe

I cannot test this change as I do not have the build environment set up. Is is intentional that this 'cloud' feature is inherited by the SGE backend? I am fixing an issue caused by a feature I do not need. I wonder if it could be placed in the epilogue for user configurations. Isn't that the purpose of epilogue and configuration?

* Use xargs -0 instead of GNU only -null