Use a termination action for the monitoring action in PAPIv2
PAPIv2 sends SIGKILL to background actions
upon instance termination.

However, we'd like to terminate monitoring action
gracefully with SIGTERM instead.

The new monitoringTerminationAction does just
that - it runs in the same pidNamespace
as the monitoringAction, but only after
all other actions (incl. delocalization) are done,
then sends SIGTERM to all processes in
that pidNamespace and waits 10 sec to finish up
(we use a fixed delay to avoid hunting for monitoring PID).