Upgrade gcloud sdk
Reasoning:

The release notes for 276, released 2020-01-14, specify that gsutil updated to version 4.47 [0]

The previous release of gsutil, 4.46, [1] was released 2019-10-31 and made it into SDK 270.0.0 on 2019-11-05.

It seems possible to me that because our "issue was fixed in the early part of December" it wouldn't have been in gsutil until 4.47 and in SDK until 276.

[0] https://cloud.google.com/sdk/docs/release-notes
[1] https://github.com/GoogleCloudPlatform/gsutil/blob/master/CHANGES.md