Refactor GcpBatchBackendSingletonActor to use Future API
The goal is to wrap the blocking calls into a Future to
avoid blocking the actor dispatcher, such a blocking
operation could cause the actor to not process any more
messages while a we are waiting for the blocking operation
to complete.

There are some unused files deleted to make the code easier
to navigate.

Warning: This likely re-introduces the race-condition where
a job status is being retrieved before it is submitted, I
have the impression that the actor is used to submit a single
job, if that's the case, we can easily leverage the actor
state to use promises for fetching the job status, such
promises wouldn't be resolved until the job gets submitted.

Note: A potential improvement from here is use promised for
calling google only once to fetch the job status.

Overall, I need to play better with the workflow to understand
which approach must be taken.

At last, there are many noisy logs useful for understanding the
flow, such logs will be eventually removed.