batch: add on-demand retry for preemption
The Batch backend currently doesn’t retry preempted jobs with on-demand
VMs, which is problematic since our goal is to save costs and avoid
failing large tasks due to a few preempted machines. This patch
reintroduces, in part, functionality removed in commit 49d675dbb95d7,
enabling the job to restart on a STANDARD VM after encountering a
VMPreemption error.